global: use llvm-libc, libc++, and ANSI C++23 for real
When I switched to Clang, I didn't know 2 things:

1. Even though I had to change a ton of crappy code to be able to
   compile it with Clang, there are two options: -fms-extensions and
   -fms-compatibility.
   -fms-extensions only enables Microsoft C++ convenient extensions
   which I use, like __declspec(property);
   -fms-compatibility "allows Clang to parse most of invalid Microsoft
   headers" and is enabled by default when compiling under MSVC as well.
2. -stdlib=libc++ is a no-op on Windows, libc++ is not even included in
   the official LLVM packages for Windows. You need to compile and then
   enable it manually.

So, in order to generate code fully compatible with ANSI C++ and use
as few MS code as possible, you need to do additional stuff.

First of all, compile libc++ AND llvm-libc from sources. Include and
link them to all projects linked statically to xrEngine.exe, which
means everything except mimalloc, oneTBB, and openal-soft.
libc++ still lacks some exception handling functions, so we either way
link with the VC runtime, but now it's only a few symbols. The rest
comes from libc++.

Then, enable a full pack of compiler options to be fully compatible with
ANSI C++. Split my forced compilation flags into 3 parts:

1. conformance_options -- flags which actually give us ANSI C++. Part of
   them was already there, one part is taken from the libfork CI script,
   and the third part I figured out myself.
2. llvm_options -- optimization flags, like -march, -flto etc.
3. libcxx_options -- overrides to use libc++ in projects instead of VC
   runtime.

llvm-libc is available on Windows only as overlay library, i.e. it
doesn't fully replace ucrt, but only a good set of C functions, like
math, mem*(), str*(), and so on. Plus I manually replaced _stricmp()
with strcasecmp() (see xrCore_platform.h). It doesn't provide own
headers in this mode, just link it statically with /force:multiple,
so that the linker won't fail on multiple references and pick
llvm-libc's symbols instead of ucrt's and that's it.

In full ANSI C++ conformance mode, several pieces of engine stopped
compiling, so adjust them. There was nothing non-trivial or potentially
ABI breaking, we should be fine.

I found good CMake build scripts for LuaJIT, use them instead of
msvcbuilt.bat as I don't trust it. Link it statically to squeeze
more link-time optimization out of it.

Finally, after statically linking the whole C++ implementation and
LuaJIT, FatLTO now links xrEngine.exe in 10 minutes instead of 40 as
it was before even with sol, so we can revert ThinLTO and use FatLTO
again. I still believe it produces more optimized code -- especially
now that I pass --lto-O3 to the linker.

Ah right, do a cleanup of xrCore headers, include only C++ headers
there, map most of custom basic helpers like _min() etc. to the std
counterparts, see XR_FUNCTION_ALIAS() for details. This should've
been done a long ago to allow the compiler to have a good time with
the engine code.

I don't know how far we are at this point to be able to switch to
LLVM-MinGW. I'm not even sure it's worth it -- I mean, Itanium ABI
(libc++abi) is most probably better than MSVC ABI and stuff, but
I'm currently happy with what I have now.

+++ ogsr_engine/xrGame/ai_space.cpp
@@ -115,7 +115,7 @@ void CAI_Space::load(LPCSTR level_name)
     R_ASSERT(crossHeader.level_guid() == levelHeader.guid(), "cross_table doesn't correspond to the AI-map");
     R_ASSERT(crossHeader.game_guid() == gameHeader.guid(), "graph doesn't correspond to the cross table");
 
-    u32 vertexCount = _max(gameHeader.vertex_count(), levelHeader.vertex_count());
+    u32 vertexCount = _max(static_cast<u32>(gameHeader.vertex_count()), levelHeader.vertex_count());
     m_graph_engine = xr_new<CGraphEngine>(vertexCount);
     R_ASSERT(current_level.guid() == levelHeader.guid(), "graph doesn't correspond to the AI-map");
